add tests and cleanup from sonar results
