+++ .github/workflows/build.yaml
@@ -5,7 +5,30 @@ on:
       - main
   pull_request:
     types: [opened, synchronize, reopened]
+
 jobs:
+  test:
+    name: Run Tests
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v4
+      - name: Set up Python
+        uses: actions/setup-python@v2
+        with:
+          python-version: '3.8'
+      - name: Install dependencies
+        run: |
+          python -m pip install --upgrade pip
+          pip install pytest pytest-mock numpy joblib
+      - name: Run tests
+        run: |
+          pytest tests/
+      - name: Upload test results
+        uses: actions/upload-artifact@v4
+        with:
+          name: test-results
+          path: tests/
+
   sonarcloud:
     name: SonarCloud
     runs-on: ubuntu-latest
